requirements.txt
streamlit
langchain
pytesseract
pillow
pdf2image
pyPDF2
huggingface_hub
python-dotenv
langchain_community
langchain_huggingface
reportlab
pymupdf